working position compaction test and equality compaction tests
